sable audio recording
There is a patch to QEMU called audio-policy, that will turn off
recording in the AC97 device based on when the toolstack turns
this on/off. dm-agent needs to fire QEMU off with the
-disable-audio-rec flag when it needs to be turned off. The problem
is this flag was not ported to the new qemu-device.c when we
moved to QEMU 1.4. This is literally just a cut/paste from the
ioemu-device.c to qemu-device.c to bring this code over.

[OXT-42]