fix-sscanf-call.patch: dm-agent can segfault depending on optimizations
domid is actually a 16bit int where the format in buff specifies an unsigned 32-bit in.
When domid is filled it will spill over into whatever is next to it on the stack. Depending
on the optimization level this could be buff itself, resulting in a segmentation fault when
buff is freed.

Bug: OXT-205

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>